add filter component and logs service
