+++ client/src/app/_services/task.service.ts
@@ -6,29 +6,18 @@ import { catchError, map, tap } from 'rxjs/operators';
 //import { config } from '../../config';
 import { environment } from '../../environments/environment';
 import { FilterService } from './filter.service';
-
-// interface FilterParams {
-//   date: {
-//     firstDay: number;
-//     lastDay: number;
-//   };
-//   units: string[];
-//   discussions: string[];
-//   projects: string[];
-// }
+import { LogsService } from './logs.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class TaskService {
-  // filterParams: FilterParams = {
-  //   date: { firstDay: 0, lastDay: 0 },
-  //   units: [],
-  //   discussions: [],
-  //   projects: [],
-  // };
-
-  constructor(private http: HttpClient, private filterService: FilterService) {}
+  serviceName: string = 'task';
+  constructor(
+    private http: HttpClient,
+    private filterService: FilterService,
+    private logsService: LogsService
+  ) {}
 
   getTaskCountByStatus(): Observable<any> {
     let config = {
@@ -40,8 +29,19 @@ export class TaskService {
     return this.http
       .get(`${environment.apiUrl}/task/countByStatus`, config)
       .pipe(
-        tap(data => this.log('fetched data from TaskCountByStatus')),
-        catchError(this.handleError('getTaskCountByStatus', []))
+        tap(data =>
+          this.logsService.log(
+            this.serviceName,
+            'fetched data from TaskCountByStatus'
+          )
+        ),
+        catchError(
+          this.logsService.handleError(
+            this.serviceName,
+            'getTaskCountByStatus',
+            []
+          )
+        )
       );
   }
 
@@ -57,8 +57,19 @@ export class TaskService {
     return this.http
       .get(`${environment.apiUrl}/task/fieldCountPerInterval`, config)
       .pipe(
-        tap(data => this.log('fetched data from getFieldCountPerInterval')),
-        catchError(this.handleError('getFieldCountPerInterval', []))
+        tap(data =>
+          this.logsService.log(
+            this.serviceName,
+            'fetched data from getFieldCountPerInterval'
+          )
+        ),
+        catchError(
+          this.logsService.handleError(
+            this.serviceName,
+            'getFieldCountPerInterval',
+            []
+          )
+        )
       );
   }
 
@@ -71,8 +82,12 @@ export class TaskService {
       },
     };
     return this.http.get(`${environment.apiUrl}/task/tagCloud`, config).pipe(
-      tap(data => this.log('fetched data from TagClouds')),
-      catchError(this.handleError('getTagClouds', []))
+      tap(data =>
+        this.logsService.log(this.serviceName, 'fetched data from TagClouds')
+      ),
+      catchError(
+        this.logsService.handleError(this.serviceName, 'getTagClouds', [])
+      )
     );
   }
 
@@ -84,8 +99,12 @@ export class TaskService {
       },
     };
     return this.http.get(`${environment.apiUrl}/task/leaderboard`, config).pipe(
-      tap(data => this.log('fetched data from Leaderboard')),
-      catchError(this.handleError('getLeaderboard', []))
+      tap(data =>
+        this.logsService.log(this.serviceName, 'fetched data from Leaderboard')
+      ),
+      catchError(
+        this.logsService.handleError(this.serviceName, 'getLeaderboard', [])
+      )
     );
   }
 
@@ -99,31 +118,12 @@ export class TaskService {
     return this.http
       .get(`${environment.apiUrl}/task/endTimeRatio`, config)
       .pipe(
-        tap(data => this.log('fetched data from TimeRates')),
-        catchError(this.handleError('getTimeRates', []))
+        tap(data =>
+          this.logsService.log(this.serviceName, 'fetched data from TimeRates')
+        ),
+        catchError(
+          this.logsService.handleError(this.serviceName, 'getTimeRates', [])
+        )
       );
-  }
-
-  /**
-   * Handle Http operation that failed.
-   * Let the app continue.
-   * @param operation - name of the operation that failed
-   * @param result - optional value to return as the observable result
-   */
-  private handleError<T>(operation = 'operation', result?: T) {
-    return (error: any): Observable<T> => {
-      // TODO: send the error to remote logging infrastructure
-      console.error(error); // log to console instead
-
-      // TODO: better job of transforming error for user consumption
-      this.log(`${operation} failed: ${error.message}`);
-
-      // Let the app keep running by returning an empty result.
-      return of(result as T);
-    };
-  }
-
-  private log(message: string) {
-    console.log(`TaskService: ${message}`);
   }
 }
