+++ client/src/app/filter/filter.component.css
@@ -1,0 +1,62 @@
+.title{
+    margin-left: 20px;
+    font-family: var(--main-font); 
+    font-size: 5vh
+  }
+  
+  .filter-inputs{
+    margin: 0 auto;
+    width: 60%;
+  }
+  
+  .project-select, .discussion-select, .unit-select, .filter-button{
+   width: 100%;
+  }
+  
+  .start-date, .end-date{
+   width: 47%;
+  }
+  
+  .start-date{
+    float: right;
+  }
+  
+  .end-date{
+    float: left;
+  }
+  
+  .entity-filter{
+    margin-right: 15%;
+    display: flex;
+  } 
+  
+  mat-expansion-panel{
+    margin: 10px 0;
+  }
+  
+  .mat-expansion-panel-header-title{
+    flex-grow: 0;
+  }
+   mat-chip{
+  } 
+  
+  mat-expansion-panel-header {
+    height: auto !important;
+    padding-top: 5px;
+    padding-bottom: 5px;
+  }
+  mat-expansion-panel-header{
+    line-height: 20px !important;
+  }
+  mat-chip-list {
+    margin: auto 0;
+  } 
+  
+  mat-form-field, mat-select, mat-option
+  {
+    text-align: right;
+    /* font-family: 'Alef' !important; */
+    font-family: var(--main-font); 
+    font-size: medium;
+  }
+
